系统权限系统完善

+++ martin-common/martin-common-core/src/main/java/com/java2e/martin/common/core/api/IErrorCode.java
@@ -12,7 +12,7 @@ public interface IErrorCode {
      *
      * @return long
      */
-    int getCode();
+    long getCode();
 
     /**
      * 获取错误信息
